+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -82,6 +82,7 @@ class RichContentEditor extends Component {
     this.copySource = registerCopySource(this.editor);
     preventWixFocusRingAccessibility();
   }
+
   componentWillMount() {
     this.updateBounds = editorBounds => {
       this.setState({ editorBounds });
@@ -418,6 +419,18 @@ class RichContentEditor extends Component {
     return modals;
   };
 
+  handleBeforeInput = () => {
+    const { handleBeforeInput } = this.props;
+    handleBeforeInput?.();
+
+    const blockType = getBlockType(this.state.editorState);
+    if (blockType === 'atomic') {
+      // fixes space click on atomic blocks deletion bug.
+      // in general, disables any input click on atomic blocks
+      return 'handled';
+    }
+  };
+
   renderEditor = () => {
     const {
       helpers,
@@ -439,7 +452,6 @@ class RichContentEditor extends Component {
       onBlur,
       onFocus,
       textAlignment,
-      handleBeforeInput,
       handleReturn,
     } = this.props;
     const { editorState } = this.state;
@@ -455,7 +467,7 @@ class RichContentEditor extends Component {
         }
         editorState={editorState}
         onChange={this.updateEditorState}
-        handleBeforeInput={handleBeforeInput}
+        handleBeforeInput={this.handleBeforeInput}
         handlePastedText={this.handlePastedText}
         plugins={this.plugins}
         blockStyleFn={blockStyleFn(theme, this.styleToClass)}